Handle TheGivingBlock API change (#28)
Context:
- Sentry caught a 400 error when querying specific organizations from TheGivingBlock API.
- After some investigation, I was able to find that the response from a GET request on `/v1/organizations/list` endpoint, contains organizations that when queried individually return a 400 error, with an error message "Organization not found". I guess that's a consistency error and bad use of REST APIs. 
- The total of organizations retrieved is 1515, on the other hand, the total of detailed organizations is 1377.

This PR solves this issue. Now, instead of throwing an error when receiving a 400, it accepts 400 errors and returns an empty object. Then when updating the cached organizations, it checks if an organization is detailed (only checks if it has a specific detailed field) and removes that organization otherwise.
I assume that it's better to remove the not detailed organizations, instead of retrieving them as well. This way, frontend won't get "confused" and we save some bandwidth in the communications.