Decrease CollectPerformance Job Interval (#37)
Our metrics API has an endpoint `/api/metrics/hour` that returns an hourly performance average. But given that the `collectPerformance` job writes the performance value every 2 hours, we are probably retrieving inconsistent values, or at least we are not retrieving what the endpoint promises. Even though we cannot guarantee that we have a new value every hour, we can enforce that if everything goes as expected, we have a new value every hour.

This PR was originated as a result of the discussion in #31 and it decreases the `collectPerformance` job interval from one hour to half an hour.