Add removeOnComplete option for a BullMQ Queue (#40)
Our Redis instance stores all jobs and their information, even though we are not using it. As a consequence, we reach the memory limit very quickly.
To avoid this behavior, this PR adds a default option ([removeOnComplete](https://api.docs.bullmq.io/interfaces/BaseJobOptions.html#removeOnComplete)) on a BullMQ Queue, which deletes the job once it is completed. In this case, I used `true` because I think there is no use for the completed jobs, but if we want to keep some jobs, we could specify that number instead of using a boolean.